+++ tests/functional_tests/apostello/test_contact_import.py
@@ -1,19 +1,12 @@
 from time import sleep
 
 import pytest
-from tests.functional_tests.utils import assert_with_timeout, click_and_wait
+from tests.functional_tests.utils import assert_with_timeout, click_and_wait, load_page
 
 from apostello import models
 
 NEW_URI = '/recipient/import/'
 DEFAULT_NUM = '+447777777771'
-
-
-def load_page(b, wt, url):
-    b.get(url)
-    assert url in b.current_url
-    sleep(wt)
-    return b
 
 
 def add_csv_data(b, wt, text):